\name{calcSteadyHcY_ICBM1_mat}
\alias{calcSteadyHcY_ICBM1_mat}
\title{calcSteadyHcY ICBM1 mat}
\description{calculte cY and h from assuming steady state and decay constants}
\usage{
calcSteadyHcY_ICBM1_mat(Ctot, iY, parms)
}
\arguments{
  \item{Ctot}{SOM C-Stock}
  \item{iY}{steady state input}
  \item{parms}{named numeric matrix with columns kY,kO and cY and h}
}

\value{parms with columns cY and h replaced}

\author{Thomas Wutzler <twutz@bgc-jena.mpg.de>}





