\name{delta2iR14C}
\alias{delta2iR14C}
\title{delta2iR14C}
\description{Convert delta unit to atomic ratio.}
\usage{
delta2iR14C(delta, iR14CStandard = c14Constants$iR14CStandard)
}
\arguments{
  \item{delta}{numeric vector of delta values}
  \item{iR14CStandard}{numeric scalar atomic ratio of the standard}
}



\author{Thomas Wutzler <twutz@bgc-jena.mpg.de>}





