\name{derivICBM1}
\alias{derivICBM1}
\title{derivICBM1}
\description{Derivative function of Basic Colimitation model.}
\usage{
derivICBM1(t, x, p)
}
\arguments{
  \item{t}{
}
  \item{x}{
}
  \item{p}{
}
}
\details{Calls forcing function p$fInput which should (see \code{\link{forcings}}). }


\author{Thomas Wutzler <twutz@bgc-jena.mpg.de>}





