\name{derivICBMDemo}
\alias{derivICBMDemo}
\title{derivICBMDemo}
\description{Derivative function of Basic Colimitation model.}
\usage{derivICBMDemo(t, x, p)}
\arguments{
  \item{t}{
}
  \item{x}{
}
  \item{p}{
}
}
\details{Calls forcing function p$fInput which should (see \code{\link{forcings}}). }


\author{Thomas Wutzler}





