\name{initStateICBM1}
\alias{initStateICBM1}
\title{initStateICBM1}
\description{Creating initial state variables for ICBM1.}
\usage{
initStateICBM1(xc12, cn = 0, iR, modMeta = modMetaICBM1(), ...)
}
\arguments{
  \item{xc12}{
}
  \item{cn}{
}
  \item{iR}{
}
  \item{modMeta}{may pass pre-calulated modMeta for efficiency.}
  \item{\dots}{
}
}

\value{Numeric matrix (nPool, nIsotopes) of state variable mass.}

\author{Thomas Wutzler <twutz@bgc-jena.mpg.de>}





