\name{initStateICBMDemo}
\alias{initStateICBMDemo}
\title{initStateICBMDemo}
\description{Creating initial state variables for ICBMDemo.}
\usage{initStateICBMDemo(xc12, cn = 0, iR, modMeta = modMetaICBMDemo(), 
    ...)}
\arguments{
  \item{xc12}{
}
  \item{cn}{
}
  \item{iR}{
}
  \item{modMeta}{may pass pre-calulated modMeta for efficiency.}
  \item{\dots}{
}
}

\value{Numeric matrix (nPool, nIsotopes) of state variable mass.}

\author{Thomas Wutzler}





