\name{meanInput}
\alias{meanInput}
\title{meanInput}
\description{provide only the mean of the litter fall}
\usage{
meanInput(input, padj)
}
\arguments{
  \item{input}{list of datastream matrices with first two columns time and value}
  \item{padj}{parameters}
}



\author{Thomas Wutzler <twutz@bgc-jena.mpg.de>}





