\name{modMetaICBMDemo}
\alias{modMetaICBMDemo}
\title{modMetaICBMDemo}
\description{Creating Meta-information for ICBMDemo.}
\usage{
modMetaICBMDemo()
}



\author{Thomas Wutzler <twutz@bgc-jena.mpg.de>}



\seealso{\code{\link{twCreateModMeta}}, \code{\link{twSetModMetaAuxGroups}}}

