\name{twSetModMetaAuxGroups}
\alias{twSetModMetaAuxGroups}
\title{twSetModMetaAuxGroups}
\description{Setting auxiliare output items in modMeta.}
\usage{
twSetModMetaAuxGroups(modMeta, auxGroupsNew)
}
\arguments{
  \item{modMeta}{
}
  \item{auxGroupsNew}{
}
}

\value{ModMeta (result of \code{\link{twCreateModMeta}}) adjusted for auxGroups, ausOutputnames, auxOutputTemplate, and nAux.}

\author{Thomas Wutzler <twutz@bgc-jena.mpg.de>}





